# FF-3205 perf: minimize copying and a bunch of other optimizations