--- a/eppo_core/benches/evaluation_details.rs
+++ b/eppo_core/benches/evaluation_details.rs
@@ -1,23 +1,25 @@
-use std::fs::File;
 use std::{collections::HashMap, sync::Arc};
 
 use chrono::Utc;
 use criterion::{black_box, criterion_group, criterion_main, Criterion, Throughput};
 
+use eppo_core::ufc::UniversalFlagConfig;
 use eppo_core::{
     eval::{get_assignment, get_assignment_details},
     Configuration, SdkMetadata,
 };
 
 fn criterion_benchmark(c: &mut Criterion) {
-    let flags =
-        serde_json::from_reader(File::open("../sdk-test-data/ufc/flags-v1.json").unwrap()).unwrap();
+    let flags = UniversalFlagConfig::from_json(
+        SdkMetadata {
+            name: "test",
+            version: "0.1.0",
+        },
+        std::fs::read("../sdk-test-data/ufc/flags-v1.json").unwrap(),
+    )
+    .unwrap();
     let configuration = Configuration::from_server_response(flags, None);
     let now = Utc::now();
-    let meta = SdkMetadata {
-        name: "test",
-        version: "0.1.0",
-    };
 
     {
         let mut group = c.benchmark_group("new-user-onboarding");
@@ -32,7 +34,6 @@ fn criterion_benchmark(c: &mut Criterion) {
                     black_box(&attributes),
                     black_box(None),
                     black_box(now),
-                    &meta,
                 )
             })
         });
@@ -45,7 +46,6 @@ fn criterion_benchmark(c: &mut Criterion) {
                     black_box(&attributes),
                     black_box(None),
                     black_box(now),
-                    &meta,
                 )
             })
         });
@@ -65,7 +65,6 @@ fn criterion_benchmark(c: &mut Criterion) {
                     black_box(&attributes),
                     black_box(None),
                     black_box(now),
-                    &meta,
                 )
             })
         });
@@ -78,7 +77,6 @@ fn criterion_benchmark(c: &mut Criterion) {
                     black_box(&attributes),
                     black_box(None),
                     black_box(now),
-                    &meta,
                 )
             })
         });
@@ -98,7 +96,6 @@ fn criterion_benchmark(c: &mut Criterion) {
                     black_box(&attributes),
                     black_box(None),
                     black_box(now),
-                    &meta,
                 )
             })
         });
@@ -111,7 +108,6 @@ fn criterion_benchmark(c: &mut Criterion) {
                     black_box(&attributes),
                     black_box(None),
                     black_box(now),
-                    &meta,
                 )
             })
         });
@@ -131,7 +127,6 @@ fn criterion_benchmark(c: &mut Criterion) {
                     black_box(&attributes),
                     black_box(None),
                     black_box(now),
-                    &meta,
                 )
             })
         });
@@ -144,7 +139,6 @@ fn criterion_benchmark(c: &mut Criterion) {
                     black_box(&attributes),
                     black_box(None),
                     black_box(now),
-                    &meta,
                 )
             })
         });
@@ -164,7 +158,6 @@ fn criterion_benchmark(c: &mut Criterion) {
                     black_box(&attributes),
                     black_box(None),
                     black_box(now),
-                    &meta,
                 )
             })
         });
@@ -177,7 +170,6 @@ fn criterion_benchmark(c: &mut Criterion) {
                     black_box(&attributes),
                     black_box(None),
                     black_box(now),
-                    &meta,
                 )
             })
         });