--- a/eppo_core/src/eval/eval_assignment.rs
+++ b/eppo_core/src/eval/eval_assignment.rs
@@ -425,7 +425,7 @@ mod tests {
         match value {
             DefaultValue::Value(v) => v,
             DefaultValue::Json(json) => {
-                ValueWire::String(serde_json::to_string(&json).unwrap().as_str().into())
+                ValueWire::String(serde_json::to_string(&json).unwrap().into())
             }
         }
     }