This PR minimizes copying that happens inside SDK. This is mostly done by using `Arc` (atomic reference counted pointer).

For strings, I used `Arc<str>` initially. However, I looked through a ton of crates and found one that is optimized for our use case (cheap cloning) — [faststr](https://docs.rs/faststr). In addition to allowing `Arc<str>`, it tries to minimize allocation when converting to/from string type and supports small string optimization (embedding strings up to 24 bytes inside the value itself, without memory allocation). I have benchmarked it and it gives even better results.

Another big realization is that most of `AssignmentEvent` object is static for the same split, so we can pre-compute events beforehand, so we don't move tons of strings around. After we select a split, we can directly get a pair of (assignment value, event).

I also ported JSON precomputing that we implemented in Go SDK (which gives a 5x boost on our json assignment microbenchmark).

As part of implementing pre-computing, I also realized that we can pre-hash salt values, so this gives a small boost for sharding.

The combined effect is a 2.5–4x boost (even on non-logging assignments). Assignment details are having a modest 20% boost.

Python in particular is enjoying these changes (I assume due to decreased allocation on Python/Rust boundary):
```
+----------------------------------+--------------+-----------------------+
| Benchmark                        | 4.0.0-before | 4.0.0-after           |
+==================================+==============+=======================+
| Evaluation                       | 11.5 us      | 2.85 us: 4.05x faster |
+----------------------------------+--------------+-----------------------+
| Evaluation (new-user-onboarding) | 2.96 us      | 261 ns: 11.31x faster |
+----------------------------------+--------------+-----------------------+
| Evaluation (numeric-one-of)      | 5.87 us      | 735 ns: 7.99x faster  |
+----------------------------------+--------------+-----------------------+
| Evaluation (regex-flag)          | 5.09 us      | 504 ns: 10.10x faster |
+----------------------------------+--------------+-----------------------+
| Geometric mean                   | (ref)        | 7.79x faster          |
+----------------------------------+--------------+-----------------------+
```