@rasendubi 2024-10-07
> The only downside to the optimizations could come up when we port this to clients, pre-computing the assignments could have a negative since for one user they will likely not trigger most flags. Something we will keep in mind for later.

This shouldn't be a huge issue as there's not a ton of work happening during the pre-compute phase (my guess is that compiling match rules is the most expensive step) and there's not much extra allocation happening as most data is shared and would be present in configuration anyway.

However, that's something I want to keep an eye on. e.g., benchmark precompute stage, so we could be more conscious about the pre-compute / evaluation tradeoff. If pre-compute becomes too huge, we may pre-compute it lazily per-flag (which also makes more sense for clients).